core: move BlockLocation to InfraExplorer.kt
