+++ core/src/test/java/fr/sncf/osrd/utils/graph/PathfindingTests.kt
@@ -3,7 +3,6 @@ package fr.sncf.osrd.utils.graph
 import com.google.common.graph.NetworkBuilder
 import fr.sncf.osrd.graph.Graph
 import fr.sncf.osrd.graph.NetworkGraphAdapter
-import fr.sncf.osrd.pathfinding.BlockLocation
 import fr.sncf.osrd.pathfinding.Pathfinding
 import fr.sncf.osrd.pathfinding.Pathfinding.EdgeLocation
 import fr.sncf.osrd.pathfinding.Pathfinding.EdgeRange
@@ -14,6 +13,7 @@ import fr.sncf.osrd.pathfinding.getTargetsOnEdges
 import fr.sncf.osrd.reporting.exceptions.ErrorType
 import fr.sncf.osrd.reporting.exceptions.OSRDError
 import fr.sncf.osrd.sim_infra.api.Block
+import fr.sncf.osrd.stdcm.infra_exploration.BlockLocation
 import fr.sncf.osrd.utils.CachedBlockMRSPBuilder
 import fr.sncf.osrd.utils.DummyInfra
 import fr.sncf.osrd.utils.graph.PathfindingTests.SimpleGraphBuilder.Edge
