+++ core/src/test/kotlin/fr/sncf/osrd/stdcm/DepartureTimeShiftTests.kt
@@ -1,9 +1,8 @@
 package fr.sncf.osrd.stdcm
 
 import com.google.common.collect.ImmutableMultimap
-import fr.sncf.osrd.pathfinding.BlockLocation
 import fr.sncf.osrd.railjson.schema.rollingstock.Comfort
-import fr.sncf.osrd.sim_infra.api.Block
+import fr.sncf.osrd.stdcm.infra_exploration.BlockLocation
 import fr.sncf.osrd.stdcm.preprocessing.OccupancySegment
 import fr.sncf.osrd.train.TestTrains
 import fr.sncf.osrd.utils.DummyInfra
@@ -30,8 +29,8 @@ class DepartureTimeShiftTests {
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
                 .setUnavailableTimes(occupancyGraph)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(secondBlock, Offset<Block>(50.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(secondBlock, Offset(50.meters))))
                 .run()!!
         val secondBlockEntryTime =
             (res.departureTime +
@@ -63,8 +62,8 @@ class DepartureTimeShiftTests {
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
                 .setUnavailableTimes(occupancyGraph)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(secondBlock, Offset<Block>(50.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(secondBlock, Offset(50.meters))))
                 .run()!!
         val secondBlockEntryTime =
             (res.departureTime +
@@ -102,8 +101,8 @@ class DepartureTimeShiftTests {
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(lastBlock, Offset<Block>(1000.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(lastBlock, Offset(1000.meters))))
                 .run()!!
         val blocks = res.trainPath.getBlocks().map { infra.getBlockName(it.value) }.toSet()
         Assertions.assertTrue(blocks.contains("b->c2"))
@@ -145,8 +144,8 @@ class DepartureTimeShiftTests {
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(lastBlock, Offset<Block>(1000.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(lastBlock, Offset(1000.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .run()!!
         val blocks = res.trainPath.getBlocks().map { infra.getBlockName(it.value) }.toSet()
@@ -193,8 +192,8 @@ class DepartureTimeShiftTests {
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(secondBlock, Offset<Block>(100.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(secondBlock, Offset(100.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .run()
         Assertions.assertNull(res)
@@ -235,8 +234,8 @@ class DepartureTimeShiftTests {
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(thirdBlock, Offset<Block>(50.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(thirdBlock, Offset(50.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .run()!!
         occupancyTest(res, occupancyGraph)
@@ -271,8 +270,8 @@ class DepartureTimeShiftTests {
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(secondBlock, Offset<Block>(50.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(secondBlock, Offset(50.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .run()!!
         occupancyTest(res, occupancyGraph)
@@ -302,8 +301,8 @@ class DepartureTimeShiftTests {
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(secondBlock, Offset<Block>(50.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(secondBlock, Offset(50.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .run()!!
         occupancyTest(res, occupancyGraph)
@@ -345,8 +344,8 @@ class DepartureTimeShiftTests {
         val res =
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(forthBlock, Offset<Block>(1.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(forthBlock, Offset(1.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .run()
         Assertions.assertNull(res)
@@ -395,8 +394,8 @@ class DepartureTimeShiftTests {
         val res =
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(forthBlock, Offset<Block>(1.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(forthBlock, Offset(1.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .run()
         Assertions.assertNull(res)
@@ -441,8 +440,8 @@ class DepartureTimeShiftTests {
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
                 .setStartTime(100.0)
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(thirdBlock, Offset<Block>(1.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(thirdBlock, Offset(1.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .run()!!
         occupancyTest(res, occupancyGraph)
@@ -471,8 +470,8 @@ class DepartureTimeShiftTests {
         val res =
             STDCMPathfindingBuilder()
                 .setInfra(infra.fullInfra())
-                .setStartLocations(setOf(BlockLocation(firstBlock, Offset<Block>(0.meters))))
-                .setEndLocations(setOf(BlockLocation(lastBlock, Offset<Block>(0.meters))))
+                .setStartLocations(setOf(BlockLocation(firstBlock, Offset(0.meters))))
+                .setEndLocations(setOf(BlockLocation(lastBlock, Offset(0.meters))))
                 .setUnavailableTimes(occupancyGraph)
                 .setTimeStep(timeStep)
                 .setMaxDepartureDelay(1000 - 2 * timeStep)
